ai-secure-review-cli/requirements.txt
# Hugging Face Integration
aiohttp>=3.8.0
requests>=2.28.0

# Code Parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-c-sharp>=0.20.0

# CLI & Interface
click>=8.0.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Utilities
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0

# Optional: for better caching
redis>=4.5.0
